Version1

+++ WpfApp1/Models/Order.cs
@@ -18,6 +18,6 @@ namespace WpfApp1.Models
         public Status status { get; set; }
         public List<Product>? products { get; set; }
         public List<Service>? services { get; set; }
-        public int Sum { get; set; }
+        public long Sum { get; set; }
     }
 }
